# openapi/public-openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "description": "Generic API response wrapper",
        "properties": {
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "ApiResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AICO Backend REST API with clean architecture",
    "title": "AICO Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scheduler/executions/acknowledge-all": {
      "post": {
        "description": "Mark all failed executions as acknowledged.\n\nQuery Parameters:\n    task_id: Optional task ID to limit acknowledgement to specific task",
        "operationId": "acknowledge_all_failed_api_v1_scheduler_executions_acknowledge_all_post",
        "parameters": [
          {
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Acknowledge All Failed",
        "tags": [
          "scheduler"
        ]
      }
    },
    "/api/v1/scheduler/executions/range": {
      "get": {
        "description": "Get all task executions within a time range",
        "operationId": "get_executions_in_range_api_v1_scheduler_executions_range_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_time",
            "required": true,
            "schema": {
              "title": "Start Time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end_time",
            "required": true,
            "schema": {
              "title": "End Time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Executions In Range Api V1 Scheduler Executions Range Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Executions In Range",
        "tags": [
          "scheduler"
        ]
      }
    },
    "/api/v1/scheduler/executions/unacknowledged-failures": {
      "get": {
        "description": "Get unacknowledged failed executions.\n\nQuery Parameters:\n    task_id: Optional task ID to filter by\n    limit: Maximum number of results (default: 100)",
        "operationId": "get_unacknowledged_failures_api_v1_scheduler_executions_unacknowledged_failures_get",
        "parameters": [
          {
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Unacknowledged Failures Api V1 Scheduler Executions Unacknowledged Failures Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Unacknowledged Failures",
        "tags": [
          "scheduler"
        ]
      }
    },
    "/api/v1/scheduler/executions/{execution_id}/acknowledge": {
      "post": {
        "description": "Mark a single failed execution as acknowledged (hide from UI).",
        "operationId": "acknowledge_execution_api_v1_scheduler_executions__execution_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Acknowledge Execution",
        "tags": [
          "scheduler"
        ]
      }
    },
    "/api/v1/scheduler/expected-runs-today": {
      "get": {
        "description": "Calculate expected number of job runs today based on cron schedules",
        "operationId": "get_expected_runs_today_api_v1_scheduler_expected_runs_today_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Expected Runs Today Api V1 Scheduler Expected Runs Today Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Expected Runs Today",
        "tags": [
          "scheduler"
        ]
      }
    },
    "/api/v1/scheduler/status": {
      "get": {
        "description": "Get scheduler status and statistics",
        "operationId": "get_scheduler_status_api_v1_scheduler_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Scheduler Status",
        "tags": [
          "scheduler"
        ]
      }
    },
...
    "/api/v1/scheduler/tasks/{task_id}/disable": {
      "post": {
        "description": "Disable a scheduled task",
        "operationId": "disable_task_api_v1_scheduler_tasks__task_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disable Task",
        "tags": [
          "scheduler"
        ]
      }
    },
    "/api/v1/scheduler/tasks/{task_id}/enable": {
      "post": {
        "description": "Enable a scheduled task",
        "operationId": "enable_task_api_v1_scheduler_tasks__task_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enable Task",
        "tags": [
          "scheduler"
        ]
      }
    },
...
  }
}